* Build with -fno-exceptions, to save some kilobytes on libST  * The part and instrument windows now use layout system, this makes them simpler to manage and look better. (a bit more work is needed)
git-svn-id: svn://localhost/sawteeth@28 32c4e4a5-1166-415a-918b-effddbf353a8

+++ Source/stBpView.cpp
@@ -47,7 +47,7 @@ void stBPView::Draw(BRect)
 	int count = _s->BreakPointCount();
 	BreakPoint *bpp = _s->GetBreakPointByIndex(0);
 	
-	int height = Bounds().Height();
+	float height = Bounds().Height();
 
 	for (int c = 0 ; c < count ; c++){
 		float start = (float)bpp->PAL * _z;
